+++ pkg/shuttle/admission_gated_allowlist_test.go
@@ -15,8 +15,9 @@ package shuttle
 
 import (
 	"context"
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 
 	"github.com/stretchr/testify/require"
 	"github.com/teradata-labs/loom/pkg/session"